fix: allow directives to not need semicolons
