+++ GSCLSP.Core/GSCLSP.Core.csproj
@@ -7,6 +7,10 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <PackageReference Include="OmniSharp.Extensions.LanguageServer" Version="0.19.9" />
+  </ItemGroup>
+
+  <ItemGroup>
     <None Update="data\iw4_builtins.json">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </None>
